Finished UI Message Feature
- the message is stored in this.state.msg. Message Component will just
  render the message.
- message is changed in Handleclick() method in Game Component

Thoughts:
I thought I needed a function to disable a board after a game, but I
don't. This is because I would reset the game after a win and have a
replay btn. I love React Library even more now. It is perfect for
writing UI, interactive applications because developers don't need to
write that much code and the code does so much.